PLUMB_0 U5: the closing refuter's findings, fixed
Six lenses, read-only, over 93dbefd0..cd21f8bf. Fourteen findings were real.
This commit is the campaign correcting itself before the cascade.

THE HALF-PROMOTION HAD TWO MORE RESIDUES, and C1's own banner named the
phrase. It promoted the copy, the map and the buffer and left behind:
  · the stale-copy CANCEL at world_gen++, still instrument-gated — with the
    witness off, a copy staged before the boundary survived it and the next
    map read the OLD world's camera;
  · cameraReadbackStaging_'s NULL CHECK in createBuffers, still gated — a
    failed allocation went unreported and the promoted copy ran against
    nothing.
And camera_pose_ was never invalidated at a world change: `valid` was set
once and never cleared, so the tide's "no eye yet" guard could not fire again
and the first sweep of a new world read the last world's camera. The
CameraPose banner states that rule; nothing enforced it. The pose is now
reset in the TEARDOWN arm.

B2's REBASE WAS INERT ON THE BOOT WORLD. world_epoch was stamped only in the
TEARDOWN arm, so a session that never changed world ran with epoch 0 — which
is exactly the long-session shape the rebase exists for. Moved to
become_destination, the ONE door every world enters by: the boot call and the
teardown call both land there.

B-G1 WAS WRONG ABOUT THE RIBBON AND MY COMMENT REPEATED IT.
GPURibbonState::time is not a birth stamp the shader differences against
signal.t_seconds; it is the SEED of ActiveRibbon::phase, which
ribbon_frame_tick accumulates and re-uploads every frame. Taking it from
gpu_seconds() is still right, for a smaller reason (a young seed suits a
float accumulator), and the comment now says the true one.

B1's CENSUS MISSED A SILENT NARROWING: mesh_gen_settled took `float& since`
and assigned ts.seconds to it — not a braced initializer, so
-Wc++11-narrowing could not see it. The parameter and all five
*_mesh_gen_since fields are double.

THE TIDE, corrected on four counts:
  · A GALLERY IS A ROW, NOT A POINT. Both arms of gallery_unwatched measured
    to the CENTRE, so a row whose far end was behind the eye read as wholly
    behind while its near end was on screen — the exact failure D1's banner
    claimed was unavailable. Both arms now measure to the nearest reach.
    TIDE_SITE_REACH = 60 is DERIVED, not chosen: the refuter priced the fan
    at ~150 wu from PAINTINGS_MAX_BY_ARCHETYPE {8,10,12,12}, but those are
    structurally unreachable — count_raw spans [1.5,4.5] and rounds to at
    most 4 — so the true maximum is sqrt(41^2+12^2)+13+3 = 58.72 wu.
  · A STEEPLY PITCHED VIEW HAS NO GROUND FACING. With the eye looking down,
    forward's horizontal component is nearly zero and a gallery dead centre
    on screen could read as behind from azimuth alone. Past
    TIDE_MAX_PITCH_FOR_CONE the distance arm decides alone.
  · BOTH FAILURE ROADS WERE SILENT. dispatch_select_gallery and
    dispatch_place_gallery can refuse AFTER the evict has already released
    the layers — so the one road where the tide is destructive-without-repair
    was the one road it said nothing about. Both now print, and the messages
    distinguish "abandoned" (nothing will retry it) from "deferred" (the
    deferred hang owns it).
  · select_gallery_for_patch NOW CLEARS dressed_at. The GalleryCenter banner
    named this trap for whoever built the tide and then the tide was built
    without paying it: a reused slot kept the previous occupant's stamp while
    active=true, which is precisely the predicate the sweep evaluates.

THE staging_reserved RATCHET IS CLOSED. dispatch_commit_gallery's host-null
arm skips commit entirely and so never released what place reserved. Once
staging_reserved passes the pool, gallery_available_staging returns 0 for
ever and no gallery is placed again — the one path that can hard-stall
coverage. REPEAT_2's G3 traced it closed on the spawn road; the tide is a
second caller that re-selects its own host, so "unreachable" stopped being
an argument.

THE INSTRUMENTS WERE WRONG IN THREE WAYS: occ= was read one write before the
caller marks the layer, so it was always one low and would have read 39/40 on
the frame the room filled; the wall road's "Hang ended" omitted the count the
commit message said both roads carried; and authored_slept_since was reset
only inside the printing branch, so a quiet stretch carried its tally forward
and the next line claimed failures from windows ago.

THREE COMMENTS ASSERTED WHAT THE CODE DOES NOT DO, the family's own listed
repeat: tick_exhibition_manifest_retry's banner still described the two-step
and the rendered clock D4 deleted; authored_wall_seconds claimed its zero is
boot when the first read is elsewhere (harmless — every use is a difference —
but stated); and D4's premise that OVERTURE_READY_TIMEOUT_S runs from boot is
false, it runs from app->world_live stamped in init_world at device-ready, so
the margin is narrower than the first telling and the banner says so.

A5 CORRECTED HALF A FALSE SENTENCE AND CARRIED THE OTHER HALF. §3 also
claimed the reconfigure recreates the depth buffer; it has not since ACQ_0
moved that to the first acquire. Both halves are now right.

TWO TOOL BUGS IN A2 ITSELF:
  · emit() WROTE THE FILE, so the two-pass wrote MIRROR_LEDGER.md twice and
    an interrupted run left the DRAFT — carrying only the eight INPUT pins —
    as the committed artifact. Composition and publication are separate acts
    now; only main() writes.
  · THE REGEX COULD NOT SEE THE ROWS IT EXISTS TO PROTECT. M7 and M1 cite
    `gallery.hpp:2165` — basename plus line, no directory — and _CITE_RE
    requires a prefix. All seven files A2 first pinned came from ONE prose
    sentence, not from the rows whose line numbers go stale. Bare basenames
    are now resolved against an index of the tree and pinned only when
    unique. The "cited-but-unpinned is 0, measured" claim in A2's message was
    a tautology and is withdrawn.
  · And ML-2w went red on the way: its regex greps this module for the
    writer's call shape, and a comment quoting that shape matched before the
    code did — a witness defeated by prose about itself. The comment no
    longer spells it.

Gates: TU PASS, G-LAW 2 GREEN, shell GREEN, sha256 PASS, score GREEN, organ
gap PASS, organ readers NO SUSPECTS. mirror_census runs twice to identical
bytes. The ledger cascade is the next commit — until it lands the pin gates
are legitimately RED, which is the campaign's own new gates working.

+++ src/cartridges/the_board/bodies/grounded.hpp
@@ -507,28 +507,28 @@ struct EntitiesState {
     // ── Arch ─────────────────────────────────────────────────────
     ActiveArch arches[Dim::MAX_ARCH_INSTANCES]{};
     bool       arch_mesh_gen_pending = false;
-    float      arch_mesh_gen_since = -1.0f;   // the settle's stamp (PANORAMA_1)
+    double      arch_mesh_gen_since = -1.0;   // PLUMB_0 refuter — double, like ts.seconds   // the settle's stamp (PANORAMA_1)
 
     // ── Column + Antenna (sibling families, shared mesh-gen flag) ─
     ActiveColumn columns[Dim::MAX_COLUMN_ONLY]{};
     ActiveColumn antennas[Dim::MAX_ANTENNA_ONLY]{};
     bool         column_mesh_gen_pending = false;  // shared by column + antenna
-    float        column_mesh_gen_since = -1.0f;
+    double        column_mesh_gen_since = -1.0;   // PLUMB_0 refuter — double, like ts.seconds
 
     // ── Palm ─────────────────────────────────────────────────────
     ActivePalm palms[Dim::MAX_PALM_INSTANCES]{};
     bool       palm_mesh_gen_pending = false;
-    float      palm_mesh_gen_since = -1.0f;
+    double      palm_mesh_gen_since = -1.0;   // PLUMB_0 refuter — double, like ts.seconds
 
     // ── Cactus ───────────────────────────────────────────────────
     ActiveCactus cacti[Dim::MAX_CACTUS_INSTANCES]{};
     bool         cactus_mesh_gen_pending = false;
-    float        cactus_mesh_gen_since = -1.0f;
+    double        cactus_mesh_gen_since = -1.0;   // PLUMB_0 refuter — double, like ts.seconds
 
     // ── Blade ────────────────────────────────────────────────────
     ActiveBlade blades[Dim::MAX_BLADE_INSTANCES]{};
     bool        blade_mesh_gen_pending = false;
-    float       blade_mesh_gen_since = -1.0f;
+    double       blade_mesh_gen_since = -1.0;   // PLUMB_0 refuter — double, like ts.seconds
 
     // ── Pyramid ──────────────────────────────────────────────────
     ActivePyramid   pyramids[Dim::MAX_PYRAMID_INSTANCES]{};
@@ -608,10 +608,13 @@ uint32_t force_spawn_portal_arch(EntitiesState& es, MachineCtx* c, wgpu::Queue& 
 // it.
 inline constexpr float MESH_GEN_SETTLE_S = 0.133f;
 
-inline bool mesh_gen_settled(bool& pending, float& since, const TimeState& ts,
+// PLUMB_0 refuter — `since` is double because ts.seconds is. It was float,
+// and `since = ts.seconds` narrowed silently: not a braced initializer, so
+// -Wc++11-narrowing could not see it and B1's census walked past it.
+inline bool mesh_gen_settled(bool& pending, double& since, const TimeState& ts,
                              const WorldState& ws) {
     if (!pending) return false;
-    if (since < 0.0f) since = ts.seconds;
+    if (since < 0.0) since = ts.seconds;
     if (!ws.world_young && ts.seconds - since < MESH_GEN_SETTLE_S) return false;
     pending = false;
     since = -1.0f;
